Flush info messages to the message log if it is popped
While warnings and errors cause the dialog to be popped and messages shown,
info messages do not. This meant that if the dialog was already popped, the new
messages were not output to it.

The info message routine now calls a routine to refresh the errors, which does
nothing unless the dialog is popped.

Fixes #570