Fix several minor ASCII Table Effects bugs
1. Auto-columns was adding a blank line to the end of the selected table, causing
problems with Move Left/Right. Now ensure that whole lines are selected, including
the final newline of the table. Fixes #578
2. If a column was attempted to be narrowed beyond its minimum, a stack trace was
output - now silently refuses to narrow further. Fixes #579
3. If converting from step format to grid format, when there was a space at the start
of the text for a table cell, the code looped forever, failing to match any of the
regexes that were attempting to manipulate from one format to the other. Leading
spaces are now accepted in the relevant regexp. Fixes #580
4. Occasional error of an uninitialised variable `$blankline` has also been fixed.