Provide fraction conversion functionality
1. Converts fractions in DP format such as 2/3 or 1-1/4 to Unicode characters
2. Three options:
a. Only convert those that have a single Unicode character representation - leave
others in DP format
b. Convert all possible to a single Unicode character and use superscripts and
subscripts for the rest.
c. Use superscripts and subscripts for all.
3. Accepts either the Fraction Slash or standard slash characters as input.
4. Always uses Fraction Slash on output with subscripts and superscripts.
5. If character preceding match is an alphabetic, then don't convert the fraction,
to avoid altering strings such as ABC1/2 or A-B-1/2, which are more likely a
code number, serial number, etc. In maths, there would normally be a space, e.g.
"A - B - 1/2" - this fraction will be converted.