Add ability to insert page labels to text file
Insert Page Makers was already in the Adjust Page Markers dialog, but this is not as
helpful as being able to insert the page labels as they would be displayed in the HTML
file. Button added to Adjust Page Markers dialog and to Txt menu where it belongs
in terms of PPing procedure.

Fixes #423