Fix small bugs relating to moving footnotes inline
1. If several footnotes were moved inline (including removing the blank lines
surrounding them) the cached begin point for searching for the next footnote was
incorrect, meaning a footnote could be skipped. Fixed by resetting the mark after
the footnote is moved
2. A warning dialog could be popped incorrectly after dealing with the last footnote.
Fixed by removing the dialog, but still exiting the loop.
3. An error could be output if a file was loaded after working on a previous file with
footnotes. The list of footnotes was cleared on file load, but not the number of
footnotes. If the user clicked "Inline" before running First Pass, it attempted to
display a footnote that it didn't have in its (empty) list.

Fixes #600