Change Mac packaging and README for Jeebies
On modern macOS versions, running unsigned binaries is not permitted.
Since Guiguts is run by volunteers, and also Jeebies is not a part of
Guiguts, it's not reasonable that we'd pay Apple to sign this program.

This PR adds an arm64 build of Jeebies, and a download script to install
Jeebies in a way that will let it run. The script detects the CPU
architecture and downloads the correct file.

The README and package.sh were also modified to reflect these changes.

commit