Improved GG icon
This file is not directly used as an icon within GG - the icon displayed for GG windows
is read from data within the code. Perl/Tk has a 32x32 pixel limit on that icon.

This icon is available if users create a Windows shortcut to link to GG or a GG folder,
and want to use the GG icon to identify it. This version has rounded corners with transparent background, unlike the previous which had a white background.