Preserve page marker positions during Auto Table conversion
When converting a text table to HTML using Auto Table, the page marker positions
all moved to the start of the table, due to the use of delete and insert. Using
replacewith instead means that page markers will be spread through the table in
the same proportions as they previously were.

Fixes #607