Remove consecutive page separators with 99% Auto
If 2 or more page separators directly follow one after the other, all but the last one
can be safely deleted when using 99% Auto.

There was an attempt in the code to do this, but it could never be triggered as the
first 5 characters of the next line were compared against a 9 character string. Also
the correct operation is to delete, not to insert a blank line.