Make hyphens mark word boundaries in Entry widgets
In Text widgets (like the main window) a hyphenated word is treated as two words
if double clicked to select a word, or if Ctrl-arrow keys are used to move forward
or backward by a word. In Entry widgets, only spaces are considered word boundaries.

Override the relevant bindings, so that hyphens (and other non-word characters)
are treated as word boundaries. Behaviour is now the same as Notepad++.

There are some inconsistencies in the way Text widgets handle words, so it did not
seem appropriate to try to emulate them. If users require it, the Text widget could be
altered to remove those in the future.