Change shortcut keys in ASCII tables dialog
The arrow keys and delete key were bound to operations in the ASCII tables dialog.
However, there are text fields in there, which require the user to be able to use the
arrow and delete keys when entering numbers.
Bind to first letter of the command instead: Next, Previous, Left, Right
Remove the Delete binding since it is not a common operation anyway.