Implement new design for Configure Page Labels
The old design created 6 widgets per page in the text, all contained in a scrolled list.
This was slow to pop and operate in books with a moderate number of pages, and
crashed or failed to display in books with a large number of pages (e.g. 2000).

New design uses a scrolled list, with the details of the selected page shown in a
section of the dialog. Mouse-1 selects a page and the details can be edited -
reflected live in the scrolled list. Also, the two most common changes have
shortcuts: Shift-Mouse-1 to cycle Arabic/Roman/ditto, and Control-Mouse-1 to
cycle Start@/+1/No Count. Balloon help text reminds the user of the shortcut
if they hover over the areas of the dialog where they can change a value.

Click and drag in the list allows scrolling of the list and mouse-2 drag does rapid
scrolling - both Tk default behaviours.

There is no noticeable delay in popping the new dialog, even with 2000+ pages.