Admit defeat on menu bar font
It is possible to rewrite the menu bar using the "old" method of a Frame
containing menubuttons, and that method does support font changing. However,
useful features such as Alt-key operation of menus and automatic posting of menus
when mouse dragging would also need supporting. The coding cost outweighs
the importance of being able to change the menu bar font.

This commit removes a minor debug variable left behind in last commit.