Improve UX when importing/exporting prep text files
The import and export dialogs now remember the previous folder used.
After import, the Save As dialog is popped, since operations that the CPer wants to
perform on the imported files, such as character counting, require the file to have
been saved. If user does not want to, they can cancel the dialog.