Avoid undefined variable error when scrolling past zero-width space
The routine that decides whether to scroll the main window or not when the cursor
moves up or down a line gets the bounding box of the current character. If that
character was a BOM/zwnbsp, the bounding box routine returned undefined
x, y, width and height.
Now, in that case, get the bounding box of the line instead of the character, and
just set the width of the box to zero. This means the height is correctly used for
scrolling purposes. Even if the BOM is the only character on the line
the line bounding box routine still returns the line height correctly.

Fixes #643