Improve default selection of path to tools
When upgrading from a previous release, if the user copied across their setting.rc,
it contained the path to each tool in the old release, which could cause problems
if the old release was subsequently deleted, or became out of date.

Now, if the tool lies under a `tools` folder, and the user's saved path looks the same
as the new suggested path under the current release, e.g. both end with
`tools\ebookmaker\ebookmaker.exe`, then use this relative path to the tool instead
of the absolute path.