Automate Auto Illus Search even more
Once a file has been chosen to use for an illustration, give the option to automatically choose the next alphabetically for the next illustration. Illos are
frequently alphabetical anyway, or can be made so. This saves lots of time choosing
the correct file for each illustration.
Also provide a skip to next/previous file to make it easy to skip over a thumbnail or
large version of a file, etc.