Fix height and width of thumbnail in image insert dialog
The thumbnail was scaled to fit in a 200x200 square, but now that the user can
set the dialog size, it meant a landscape thumbnail sometimes caused excess space
in earlier widgets and/or a portrait thumbnail was clipped off the bottom. This was
exacerbated by the fact that when the dialog is initially popped there is no image
to display in the thumbnail space.
Pre-allocating a 200x200 space eliminates these problems.