Resolve focus issues with compose dialog
If the compose dialog is already popped, and an attempt to re-pop it is made, it
can end up with the focus, but the focus is not on any of its three widgets.

To avoid this, just destroy the dialog if it is popped and recreate it. It is a rare
occurrence, and just causes a brief flash of the dialog.

This may resolve the issue referred to in #654, but is probably unrelated.