Permit case where block markup ends then restarts on the next line
The manual warns that close block markup should have a blank line
after it, and open block markup should have a blank line before. However, it is quite
easy to end up with close followed by open, e.g. if a table spans a page boundary.
The formatters will have marked each page as a table and any blank line could be
lost during page separator removal.

Although the comment said "find end of paragraph or end of markup", it actually
only searched for end of paragraph, which did not find the closing markup when
it had no blank line after it. Now fixed.

Also, if the whole file was selected with Ctrl+a, the test for the end was faulty and
ended up looping through the whole file twice. Also fixed.