Add betacode ordering for Greek letters to Compose sequences
It is not possible for Compose = a to produce alpha, but Compose = a / to produce
alpha with accent, because the former pre-empts the latter.
Therefore, in order to support betacode ordering of keystrokes, it is necessary to
either insist on pressing Enter for every Greek compose sequence, even the plain
letters, or to have a different introductory character.
This commit does the latter: if introduced with hyphen/minus then betacode
ordering is used and requires Enter/OK when complete; if introduced with equals
then as Greek works as previously.
To free up -L for lambda, it was necessary to change the sequence for the pound
sterling symbol from -L to /L  or \L.

Fixes #664