Tell user how to install and configure kindlegen if needed
DP wiki page covering the installation of helper apps on Windows has been updated.
User is directed to look there in INSTALL file. Mac users are given information in
the INSTALL file since the DP wiki Mac page is out of date.

Default folder to find kindlegen in the code is now correct for Windows and Mac
systems.