Cope with block markup at end of footnote
Block markup at end of footnote may not have a blank line after it. Although this is
a requirement according to the manual, it is better not to require it.

The end position of the block wasn't being placed correctly for this circumstance.
Now corrected to fix the second half of #688.

#685 already partially fixed #688