Make word frequency letter shortcuts handle æ ligature better
In word frequency, the æ ligature is sorted as though expanded into the two letters ae.
The "a" key shortcut was not handling this correctly, since the logic did not deal with
the ligature being expanded to two letters. Now just the first letter is used. This
means it now works whether there are words beginning with a that are alphabetically
before æ or if the first word beginning with "a" actually begins with æ.