Highlight characters in WF that are not in an enabled character suite
1. List of available character suites is available via the File->CP menu.
2. User can enable/disable them (except Basic Latin).
3. Check button in File->CP menu controls highlighting of any characters in the WF
Character Count list that are not in an enabled charsuite.
4. Control-clicking a highlighted line either pops a dialog asking if the user wants to
enable the first charsuite alphabetically that contains the character, or if there is no
such charsuite a warning dialog is popped
5. A bug (from 1.0.25) where the user can control-click in a word listing in the WF
dialog to display 1st harmonics has been fixed (missing argument to routine)
6. An undocumented broken feature (again since 1.0.25) to add a word to the project
dictionary with control-mouse 2 or 3, has been removed.