Optimise status bar updating
The status bar is updated every 200 milliseconds. The update routine can take more
than 200ms to run under some circumstances. The slow part is finding the current
page number when in a large file without page numbers and lots of marks have
been inserted by a checking tool. The code takes time going through those marks
looking for the previous/next page mark. As a result there is no time for anything
else to happen and the program becomes unusably slow.

Fix is to remember the current line/column and only search to find what page
we are on if the line/column have changed since the previous update.