Avoid leaving page boundaries within a word
When lines are joined during Page Separator Fixup, the page boundary could end up
mid-word. When HTML was generated, pagenum spans were output where the page
boundary mark was, even if it was mid-word.

Now, Page Separator Fixup and HTML autogeneration both shift the location
forward to find a position that is not mid-word.

Fixes #733