Avoid combining/skipping page numbers unless no text between them
Previous check for whether next page number was at least one line ahead of the
current one was unsafe when there is very little text on a page, e.g. with long,
continued footnotes, there is sometimes just one line of main body text on a
page. If this line was shorter than the previous one, it could lead to the page
being treated as a blank one, so the page numbers would be combined or
skipped. This has happened since 1.0.25 at least, but with the new feature
where consecutive page labels can be skipped rather than combined, the
consequences are more significant.

Now checks if there are any word characters between the page breaks or not.