Search correctly in WF when word has underscore before or after
In order to force "whole word" searching, code used `\b` before and after.

However `\b` in Perl does not include the underscore character so this failed
when word had underscore adjacent to it (used for italic markup). Solution is
to use a negative lookbehind and lookahead, only matching alphanumeric
characters not underscores.

Fixes #743