Don't store width & height for bookloupe view options dialog
There is no reason to resize the dialog as it has fixed labels in it. Having width
and height stored means that a font change can cause some of the dialog to
be hidden.

Change the entry from the geometryhash to the positionhash - there is already
code to handle such a change when the user upgrades to a new version in
FileMenu.pm

Fixes #746