Avoid advancing page markers into HTML tags
Page markers were advanced to the next whitespace to avoid them appearing
mid-word (#738).

However in HTML, it is possible to have a `<br />` or `<span...` immediately following a word, and the page marker could be advanced into it.

Avoid this by advancing to the next whitespace or `<` when inserting HTML
page number spans

Fixes #756 .