Stop using charset ascii when autogenerating HTML
If a file had no non-ascii characters, and the default option to treat all files
as utf-8 was turned off, the HTML header set the charset to ascii.
Although this is valid, some tools whine about it and suggest us-ascii instead.

Since for an ascii file, the charset iso-8859-1 is equally valid, use that when
ascii would have been used previously

Fixes #755