Add auto-correction features to error check dialog
Add features to process queries in the error check dialog, initially for Jeebies
and Load Checkfile (to supoprt OCRfixr output)

For Jeebies, the change will be to swap he/be.
For Load Checkfile, if the text is `Suggest 'X' for 'Y'` then replace Y with X at
marked position.

1. Ctrl+Mouse-1 makes the change suggested by the query
2. Ctrl+Mouse-2/3 makes the change and removes the query (as Mouse-2/3
currently does)
3. Ctrl+Shift+Mouse-2/3 removes all queries matching the clicked on, i.e.
to get rid of multiple wrong identical suggestions

Fixes #753