Fix several Basic Fixup bugs relating to block markup
1. `/F` frontmatter markup now treated as block markup, like `/X`, so Fixup
checks are skipped by default
2. Bug in `/X` handling fixed - the closing `X/` wasn't previously detected
causing checks to be omitted for the rest of the file. Fixed by putting the
block types to be checked into a variable.
3. `/P` markup for poetry was (and is) not included in the block types for
which checks are skipped by default. Since one of the checks is reducing
multiple spaces to a single space, this broke poetry line numbers which
are represented as multiple spaces then a number at the end of the line.
Now, if in `\P` markup, temporarily remove then restore any poem line
number in the space compressing code.