Fix error when chosen scannos directory is guiguts install directory
Although this seems unlikely at first glance, when Set File Paths is used to
select a scannos directory, the default is for it to open at the parent of the
current scannos directory, which is the main install directory. Clicking OK
at this point used to give an error due to a substr argument being out of
range when the code was trying to convert to a path relative to the install
directory.

Now checks the new path is long enough to for a legal substr call.