Don't store duplicate errors from bookloupe
A duplicate error (i.e. same error at same line and column) message from
bookloupe causes an error later when the user attempts to delete the lines
as they process them. This is because the errors are used as the keys into a
hash containing the position in the text to find the error. Therefore when the
first occurrence of the error is deleted by the user, the hash entry is cleared,
and then the second occurrence can't find an entry in the hash.

There is no benefit in exact duplicate errors anyway, since the user won't know
what the second one means.

An example of a file causing bookloupe to give a duplicate error is
```
"?
"D
```
which causes two errors warning about bad quote spacing, both referencing
the quotes in the first column of the first line.

Fixes #770