Output line number in unused CSS message from pphtml
Minimal changes required to store and track the line number where CSS is
defined in order to be able to output it later. I do not want to make substantial
edits to pphtml because it may be used separately from GG, because its output
is very familiar to experienced PPers and because it's a relatively simple
peripheral program that is easy to understand with its current structure, and
best not meddled with.

Fixes #754