Fix minor character insert quibbles
1. Small border added around common character insert buttons to make it
possible to click the border to raise the dialog to the top - useful since it is a
dialog that is likely to be left on-screen while other dialogs are used. Fixes #766
2. Check that the widget into which the inserted character will be placed still
exists - slightly more likely this will get triggered now that more widgets are
able to accept the inserted character. Will still be rare, but behaviour will be
appropriate - if focus is not on a widget that can accept special character input,
nothing will be inserted
3. Increase the width of the Surround and Flood dialogs - they are so narrow
that it is hard/impossible depending on your font sizes to pick up the banner to
reposition them.  Fixes #765
4. Allow the Surround and Flood dialog to accept special character input from
Common Characters, Compose Character, etc.  Fixes #759