Clarify match numbers displayed by spellchecker
When spell check finds a misspelled word, it displays up to 4 numbers which
detail variants on the word: exact matches, case variants, possessive variants
(add or remove 's), hyphen variants.

Since there can be 1, 2, 3 or 4 numbers displayed, and the meaning of the
numbers varies depending on how many are displayed, it's not easy for the
user to know/remember what each one means. Previously only hyphens were
labelled.

Now, output an extra word to describe the type of match, eg.
instead of
`3, 2, 1, 4 hyphens in text.`
output
`3 exact, 2 case, 1 possessive, 4 hyphen in text.`

This is only intended as a reminder; the manual will give more detailed
explanations of the meaning of those words, e.g. whether case is all case
variants or just capitalised first letter, that possessive means either adding
or removing `'s`, and that hyphen means the word is used in a longer
hyphenated word, not that it appears solo in a hyphenated form,  e.g.
`tomorrow` would flag `tomorrow-people`, not `to-morrow`.

There was also a bug previously where the first spelling found did not display
the additional types of match, only exact matches. This was due to partly
duplicated code in subs spellcheckfirst and spellchecknext. Changed to call
spellchecknext at end of spellcheckfirst and removed duplicated code.

Fixes #760