Add more options to Check Orphaned Brackets
Various markups are already checked, such as `/* */` and `/$ $/`.
Add `/X X/`, `/F F/`, `/I I/`, `/L L/`

Also, the printable_brackets routine, that converts the regex used for searching
into a more user friendly form for display on a button and label, had not been
updated to cope with `/[Pp]|[Pp]/`. Since this is also the format used by the new
options, it has now been updated to convert this to `/P P/`, etc.

Fixes #772