Improve CSS used for indexes
As described in #774, index lines should really be indented when they wrap.
This CSS means the wrapped lines are 2em further indented from the first line.
Each sublevel of the index continues to indent by 1em from the previous
sublevel, so wrapped lines are not confused with a following sublevel.

Fixes #774