Improve quote/bracket highlighting
Previously, it was necessary to change the text between two quotes or brackets
for them to be highlighted by the Preferences->Enable Quotes Highlighting
operation. There was also no easy way to remove the highlighting apart from
moving the cursor somewhere in the text where it wasn't between quotes or
brackets and modifying the text again. The code used a built-in feature of
the Text widget.

Now, combine the built-in Text widget feature with the control method used
for scanno highlighting, i.e. set a repeating callback to the highlight routine.

Also, adjust the default behaviour to make it check for matching quotes across
multiple lines as it does for brackets, instead of single line as before.
Also add curly quote support.

Fixes #780