Fix long-standing bug in Search Orphaned Brackets
If markup such as `/*` or `/P` was used and never closed, it was not flagged, but
an error message was output to the command window (at least 1.0.25 onwards).

This was due to the array element for the closing markup being accessed when
it wasn't defined.

Also added `/i` markup to the previous fix in #793.