Correct handling of bookloupe errors
The fix in #791, to avoid copies of the errors hash being inadvertently retained
by Perl due to separate subroutine calls and recursion, was wrong.

Instead of making the hash a local variable passed by argument to the subs
that need it, return it to being a global within the package, which does not
need double colon prefix when using it.

Fixes #795