Fix long-standing right-click bug
If you right-clicked 100 times, you would hit a "deep recursion" error. If you persisted, the program would crash, presumably once the recursion got too
deep. The fault was that the code called `Popup` instead of `post`. The former
does not return, leading to an extra level of recursion each time the right-click
menu was used. The `post` method is the correct one for this case.

Fixes #801