Make right-click menu contain just a few useful options
Instead of right-click bringing up the whole menu structure, which makes it
superfluous, align it with other programs by popping a small menu of useful
options, namely Cut, Copy, Paste and Bookmark options. Also, make it available
on all platforms - previous vague comment about trapping odd error may
refer to the long-standing bug fixed by #802.

In addition, previously the location of the right-click was ignored, meaning that
Paste and other options behaved in an unexpected manner. Now, the location
of the right-click is used for Pasting and Bookmarks.

Note that (fortuitously) if you right click over selected text and paste, it replaces
the selected text, but if you right click over a part that isn't selected, the
selected part is not replaced. I think this is desirable behaviour.

Fixes #800