Preserve white background for Windows check boxes
The logic was failing due to lack of brackets to enforce operation precedence
with the ternary operator, meaning that although on Windows systems check
boxes should remain white, they did not.
Note that this code was executed even if the Color Choose dialog was
cancelled, which is fine because with the correction it will have no effect, as
was probably the original coder's intent, and is the structure used in similar
Color Choose code nearby.