Make searches wrap during curly quote processing
To avoid the user thinking prematurely that they have done all straight to
curly quote processing, make the searches for "@" and straight single quotes
wrap at the end of the file. Now if the user gets a bell and icon flash, it's
because there are no more "@" or straight single quotes left in the file.

Fixes #816