Make HTML conversion of footnotes and sidenotes more robust
Use of square brackets within sidenotes or rare footnote constructions could
lead to error messages and abandonment of HTML conversion.

Check that footnote data is all defined before trying to use it, and improve
the handling of square brackets within sidenotes, to avoid system errors that
did not help the user anyway.

Fixes #824