Stop jeebies hanging if a word has more than 50 characters
The maximum word length was set to 50 in the code, but a "word" could be
longer than this - actual example was "of" followed by 49 leader dots.

Ideally, jeebies code could/should be altered to make it robust, even if the
maximum word length is exceeded. However, given that line lengths in text
files are less than 80, and the situation clearly only arises extremely rarely,
changing the maximum to 100 seems fine for now. The whole of jeebies was
written long ago, and would probably use different data and code structures
if written today.

Fixes #826