Ensure replacement uses the same regex options as search
When replacing `$n` with a bracketed section of the regex, the "s" flag was
not used on the regex search. This flag causes "." to include newlines. Since
the flag is used when the user is searching for the regex in the first place,
it needs to also be used when doing the replacement. Otherwise, the user
does a Search and finds a match, but when they try to replace, they just get
`$1` left in their file.

Note this should have no effect on the commonly used `(.|\n)` construction.

Fixes #829