Fix index text wrapping bug after page marker
On the first line of each page, it was possible for the temporary character used
to replace the page markers to fool the wrapping code into starting the wrap
from the second character on the line, leaving the first character stranded at
the left margin.

Each chunk that gets wrapped should begin at the start of the line.

Fixes #834