Avoid misidentifying indent markup as footnote anchor
The markup `/*[4]` is used to set the indent for a block to be 4.
This could sometimes be mistaken for a footnote anchor if the numbering
happened to match, e.g. it occurred between the actual `[4]` anchor and the
footnote itself.

Fixed by using a regex instead of a text match to make sure the potential
anchor is not preceded by the relevant markups `/* /# /I /i`

Also adjusted wording in Footnote Check dialog to better match the reason
for the warning.

Fixes #846