Update to ebookmaker 0.11.26
Ebookmaker 0.11.26 includes a command line option to indicate the location
of a tidy config file. In the config file, tidy is instructed to convert named
HTML entities to numeric. This is required for ebookmaker's conversion to
HTML5.
The config file is included in the ebm_builder release which package.sh grabs
during the release process, and is placed in the ebookmaker tools folder.

Fixes #856