Bundle the Nu HTML Checker
The Nu HTML checker is use by W3C online checker if an HTML5 file is being
validated. Bundle this with Guiguts in place of the old onsgmls.exe. The .jar
file is quite large, but in time, the old CSS jar file might be retired, because
the Nu checker also checks CSS, but retaining the existing separate CSS
checker for now.

Replace the previous W3C HTML validation with two options to use the Nu
checker. The first checks as HTML, the second as XHTML. Although it had
been postulated that the XHTML test was sufficient, there are messages,
e.g. warning about use of "name" attribute, that only occur in the HTML check.
The XHTML check is necessary to catch unclosed elements, e.g. `<hr>`

Added new XHTML test, and adjusted several other tests due to the
input test html file having a different (HTML5) header, which affected
line numbers.