Fix out-of-date list of required perl modules in INSTALL.md
The list in INSTALL.md had become out of sync with the list of modules
installed by the helper script `install_cpan_modules.pl`.

Rectify this by removing the list from INSTALL.md and referring the user
to the list in the script if they prefer to install them manually.