Convert HTML4 header.txt to HTML5 (#891)
If user has copied a customized header.txt from a previous release, it will have
an (X)HTML4 header. As this is read, convert the few lines necessary to make
it valid HTML5.

This allows the user to upgrade fairly painlessly from a 1.3 (HTML4) release to
a 1.4 (HTML5) release.

Also, beef up the upgrade notes a bit to clarify procedure for retaining any
header.txt customizations.