Convert sub/superscripts correctly in block markup
During HTML generation, the routines to convert sub/superscripts were called
for each line of the file near the beginning of the conversion.
They edited the file directly to add the sub/sup markup.
The calling routine ignored this change and continued to work with the line
it had extracted from the file before the sub/sup conversion. This meant that
if further work was done on the line, e.g. adding a span to indent it for a table,
the sub/sup edit was lost.

Fixed by returning the modified line to the calling routine - it didn't seem
worth trying to restructure, since that would open a can of worms regarding
the structure of the HTML conversion with its massive outer loop.

This has been a bug since 1.0.25 at least; very surprising it hasn't been
reported/fixed before. Possibly HTML files have been posted with the
`^` or `_` markup still in where sub/sups occur inside block markup, since
it would be easily overlooked by the PPer post-conversion.