Use CRLF line endings when saving main text file
DP text files use CRLF line endings, and some tools expect this.
Guiguts already forced this for Windows when saving the main text file.
Mac/Linux users have said it would be easier if it did the same for them.

A variety of line endings are accepted when loading a file.

Fixes #894