Strip all rewrap markup from input to bookloupe
Bookloupe only knows about the 3 oldest rewrap markup types, which it
does by stripping them from each line before it checks it.
Since we are not editing bookloupe code, in order to make bookloupe ignore
the newer rewrap markup types, we need to strip them from the file before
sending it to bookloupe.

Only strip if markup is on its own on the line. Note that opening markup can
be followed by `[` to specify margins.

Fixes #901