Cope better with the last-used directory being deleted on Macs
If the directory last used to load/save a file in is deleted, the next time the
File Open dialog was popped, an error was displayed on Macs (and possibly
Linux). On Windows, it just defaulted to open the GG install directory if
the attempted directory didn't exist.

This commit makes all platforms emulate that behaviour.