Ensure ids used for inserted images are unique
If an image was used several times in an HTML file, e.g. a decorative detail at
chapter headings, it was given the same id each time, based on the filename.

Now, if it finds the id has been used already, it appends `_2`, `_3`, etc., until
it finds one that hasn't been used.

Fixes #914