Correctly extract authors' names containing the letter "F"
An incorrect regex, in the wrong place, and not really necessary at all, meant
that authors' names containing the letter "F" were not extracted from the
text file to pre-populate the HTML generation dialog. This only happened
when the author's name was on the same line as the word "By", e.g.
"By Foo Bar".

The regex was trying to detect closing `/F` or `/f`, which should have been
above the containing `if` block anyway, and isn't needed as there are other
checks that the searching process stops in reasonable time. The fix is
therefore to use remove the check completely.

The regex should have been `^/[Ff]` but was `[/[Ff]` which would match either
a `/`, `[`, `F` or `f`.

The bug has probably been there "for ever", but the pre-population is just a
convenience, not an essential for HTML generation, and would only occur
rarely, so has escaped notice until reported by an eagle-eyed user.