Remove "View in Browser" from HTML generation dialog
It is confusing to have "View in Browser" adjacent to "Autogenerate HTML"
because if autogenerate then immediately press "View in Browser", it doesn't
show the generated HTML since it hasn't been saved yet.

Since "View in Browser" is hardwired into the Custom Menu, there's no need
to have it in this dialog anyway.

Fixes #915