Use manual pages for shortcuts and regex help
Help is generally supplied via the wiki manual pages, but the shortcut keys
and regex reference charts were historically bundled with the release and
displayed in their own dialogs. Having this information in 2 places is
undesirable for maintenance, so remove the existing text files and link
the Help Menu options to the manual instead.

Fixes #