Add Combining Characters to Compose Key Sequences
PPers are being encouraged to use combining characters when a precomposed
Unicode character does not exist.
Add the more common characters to the available Compose Sequences.
Each is introduced with a plus sign, e.g. `+~` for combining tilde, `+-` for
combining macron, etc.

Also, add ability to provide a description in the Help dialog, to make it
clearer which character can be created with the Compose Sequence.