Fix blockquote footnote clash
The problem was the footnote code assumes it is the only thing that puts </p></div> (to end a footnote). However, if you use CSS blockquote markup (the default), that also puts </p></div>. All I've done is make it use the non-CSS blockquote markup <blockquote> initially. Then after the footnotes are finished, convert them to using divs if that's what the user has selected.
I think this is safest, and fairly easy to understand, given the delicate state
of the footnote code.

Fixes #949