Set kindlegen path to an absolute path
The default path to kindlegen on startup (without setting.rc) is
"Applications/Kindle Previewer 3.app/Contents/lib/fc/bin/kindlegen"

Without a leading '/' or './' or '../' Guiguts will consider this
a relative path. Unless the user happens to start guiguts while
in the root directory, this will never work.

This patch adjusts the default value for $kindleviewercommand
to include the leading '/'