Add --home to specify location for persistent data
Adds a --home flag, taking a directory as an argument, which
specifies a directory for storing persistent user data. At
this time, three things would be stored in this directory:

 - setting.rc
 - header.txt
 - labels_{lang}.rc

In the case of a corrupted setting.rc, then this location is
also where setting.err might be written.

This flag is optional, the default behavior is preserved.

Changes argument parsing to use Getopt::Long. Options --home
and --runtests are parsed with Getopt::Long. If a filename is
specified, that is still handled by reading @ARGV.

Note that this means a new arg "--runtests" replaces the older
form "runtests".

Also adds helper functions to return various paths:
 - header.txt
 - headerdefault.txt
 - labels_{lang}.default.rc
 - labels_{lang}.rc
 - setting.rc

Some of the helpers are to simplify things with the variable
location (whether --home is in use or not), some are just for
the sake of convenience.

Fixes #954